+++ src/test/2023-07/Bao_exp.sol
@@ -20,14 +20,22 @@ import "./../interface.sol";
 // https://blog.hundred.finance/15-04-23-hundred-finance-hack-post-mortem-d895b618cf33
 
 interface IbSTBL is IERC20 {
-    function joinPool(uint256) external;
-    function exitPool(uint256) external;
+    function joinPool(
+        uint256
+    ) external;
+    function exitPool(
+        uint256
+    ) external;
 }
 
 interface IbdbSTBL is IERC20 {
     function mint(uint256, bool) external;
-    function redeem(uint256 redeemTokens) external;
-    function redeemUnderlying(uint256 redeemAmount) external;
+    function redeem(
+        uint256 redeemTokens
+    ) external;
+    function redeemUnderlying(
+        uint256 redeemAmount
+    ) external;
 }
 
 contract ContractTest is Test {
@@ -85,7 +93,7 @@ contract ContractTest is Test {
         uint256[] memory amounts = new uint256[](2);
         amounts[0] = 17_550_000 * 1e6;
         amounts[1] = 17_510_000 * 1e18;
-        uint256[] memory modes = new uint[](2);
+        uint256[] memory modes = new uint256[](2);
         modes[0] = 0;
         modes[1] = 0;
         AaveV2.flashLoan(address(this), assets, amounts, modes, address(this), "", 0);
