+++ src/test/2023-07/Libertify_exp.sol
@@ -70,7 +70,7 @@ contract ContractTest is Test {
         assets[0] = address(USDT);
         uint256[] memory amounts = new uint256[](1);
         amounts[0] = 5_000_000 * 1e6;
-        uint256[] memory modes = new uint[](1);
+        uint256[] memory modes = new uint256[](1);
         modes[0] = 0;
         aaveV2.flashLoan(address(this), assets, amounts, modes, address(this), "", 0);
         aaveV2.flashLoan(address(this), assets, amounts, modes, address(this), "", 0);
